% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacoCode.R
\name{plotAUCConsistency}
\alias{plotAUCConsistency}
\title{Function to plot the AUC consistency between the different projects}
\usage{
plotAUCConsistency(data, drugName)
}
\arguments{
\item{data}{A data frame of summarized pharmacogenomic data with column names
"cellLine", "drug", "ic50_CCLE", "auc_CCLE","ic50_GDSC", "auc_GDSC".}

\item{drugName}{A character vector containing the name of the drug to evaluate.
This value has to be contain in the "drug" column from the "data" data frame.}
}
\value{
A ggplot2 object
}
\description{
\code{plotAUCConsistency} uses ggplot2 functions to plot a scatterplot of
AUC measurements in the CCLE project as a function of the AUC measurements of
the GDSC project. Based on this scatterplot, one can evaluate the consistency of
these measurements. If the data points are close to the f(x) = x line, it
would mean that the measurements are consistent between the different projects.
}
\examples{
library(PharmPlotter)
data("summarizedData", package="PharmPlotter")
pl <- plotAUCConsistency( summarizedData, "AZD0530" )
}
